fix(bot): widen audit_logs.telegram_user_id to BIGINT
Telegram IDs routinely exceed int4 (2^31), so audited manager actions
(approve/block/reject, role and specialization changes) failed with
NumericValueOutOfRange and rolled back the whole action. Migration 004
was recorded as applied on prod (alembic at 009) but the column stayed
int4 — table created from the ORM model via create_all and versions
stamped without the ALTER running.

Adds migration 010 to re-apply the widening idempotently and switches the
AuditLog model to BigInteger so create_all stops recreating int4. Prod DB
already altered live. Documented as MGR-01 in docs/bugs-2026-05-28.md.

+++ uk_management_bot/database/models/audit.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, DateTime, Text, ForeignKey, JSON
+from sqlalchemy import Column, Integer, BigInteger, String, DateTime, Text, ForeignKey, JSON
 from sqlalchemy.sql import func
 from sqlalchemy.orm import relationship
 from uk_management_bot.database.session import Base
@@ -13,7 +13,8 @@ class AuditLog(Base):
     user = relationship("User")
     
     # Telegram ID пользователя (для сохранения логов после удаления пользователя)
-    telegram_user_id = Column(Integer, nullable=True, index=True)
+    # BigInteger: современные Telegram ID превышают int4 (2^31)
+    telegram_user_id = Column(BigInteger, nullable=True, index=True)
     
     # Тип действия
     action = Column(String(100), nullable=False)
